more tests; finalize day7
