+++ src/refsols/mod.rs
@@ -9,3 +9,5 @@ pub mod refsol_day4;
 pub mod refsol_day5;
 
 pub mod refsol_day6;
+
+pub mod refsol_day7;
